Update DropdownMenu for changes in core
