+++ packages/patternfly-4/react-core/package.json
@@ -47,7 +47,7 @@
     "@patternfly/react-tokens": "^1.0.0"
   },
   "devDependencies": {
-    "@patternfly/patternfly-next": "1.0.61",
+    "@patternfly/patternfly-next": "1.0.63",
     "css": "^2.2.3",
     "fs-extra": "^6.0.1",
     "glob": "^7.1.2",
