+++ packages/patternfly-4/react-core/src/components/Dropdown/DropdownMenu.js
@@ -3,6 +3,7 @@ import styles from '@patternfly/patternfly-next/components/Dropdown/dropdown.css
 import { css } from '@patternfly/react-styles';
 import PropTypes from 'prop-types';
 import { componentShape } from '../../internal/componentShape';
+import { DropdownPosition } from './dropdownConstants';
 import FocusTrap from 'focus-trap-react';
 
 const propTypes = {
@@ -13,28 +14,47 @@ const propTypes = {
   /** Flag to indicate if menu is opened */
   isOpen: PropTypes.bool,
   /** Indicates which component will be used as dropdown menu */
-  component: componentShape
+  component: componentShape,
+  /** Indicates where menu will be alligned horizontally */
+  position: PropTypes.oneOf(Object.keys(DropdownPosition))
 };
 
 const defaultProps = {
   children: null,
   className: '',
   isOpen: true,
+  position: DropdownPosition.left,
   component: 'ul'
 };
 
-const DropdownMenu = ({ className, isOpen, children, component: Component, ...props }) => {
+const DropdownMenu = ({ className, isOpen, position, children, component: Component, ...props }) => {
   let menu = null;
   if (Component === 'div') {
     menu = (
-      <Component {...props} className={css(styles.dropdownMenu, className)} hidden={!isOpen}>
+      <Component
+        {...props}
+        className={css(
+          styles.dropdownMenu,
+          position === DropdownPosition.right && styles.modifiers.alignRight,
+          className
+        )}
+        hidden={!isOpen}
+      >
         {children}
       </Component>
     );
   } else if (Component === 'ul') {
     menu = (
-      <FocusTrap className={css(styles.dropdownMenu, className)}>
-        <Component {...props} hidden={!isOpen}>
+      <FocusTrap>
+        <Component
+          {...props}
+          className={css(
+            styles.dropdownMenu,
+            position === DropdownPosition.right && styles.modifiers.alignRight,
+            className
+          )}
+          hidden={!isOpen}
+        >
           {children}
         </Component>
       </FocusTrap>
