+++ packages/patternfly-4/react-core/src/components/Dropdown/dropdownConstants.js
@@ -1,0 +1,9 @@
+export const DropdownPosition = {
+  right: 'right',
+  left: 'left'
+};
+
+export const DropdownDirection = {
+  up: 'up',
+  down: 'down'
+};
