+++ packages/patternfly-4/react-core/src/components/Dropdown/index.d.ts
@@ -1,9 +1,6 @@
-export {
-  default as Dropdown,
-  DropdownPosition,
-  DropdownDirection,
-  DropdownProps
-} from './Dropdown';
+export { default as Dropdown, DropdownProps } from './Dropdown';
+
+export { DropdownDirection, DropdownPosition } from './dropdownConstants';
 export { default as DropdownItem, ItemProps } from './Item';
 export { default as DropdownSeparator, SeparatorProps } from './Separator';
 export { default as KebabToggle, KebabProps } from './KebabToggle';
