+++ packages/patternfly-4/react-core/src/components/Dropdown/index.js
@@ -1,4 +1,5 @@
-export { default as Dropdown, DropdownPosition, DropdownDirection } from './Dropdown';
+export { default as Dropdown } from './Dropdown';
+export { DropdownDirection, DropdownPosition } from './dropdownConstants';
 export { default as DropdownItem } from './Item';
 export { default as DropdownSeparator } from './Separator';
 export { default as KebabToggle } from './KebabToggle';
