+++ packages/react-icons/package.json
@@ -33,7 +33,7 @@
   },
   "devDependencies": {
     "@fortawesome/free-solid-svg-icons": "^5.3.1",
-    "@patternfly/patternfly-next": "1.0.61",
+    "@patternfly/patternfly-next": "1.0.63",
     "fs-extra": "^6.0.1",
     "glob": "^7.1.2",
     "node-plop": "^0.15.0",
